Extract projectCountry field
