+++ controllers/apply/awards-for-all/fields.js
@@ -25,12 +25,12 @@ const {
     FREE_TEXT_MAXLENGTH
 } = require('./constants');
 
-const countriesFor = require('./lib/countries');
 const locationOptions = require('../lib/location-options');
 
 const fieldContactLanguagePreference = require('./fields/contact-language-preference');
 const fieldOrganisationStartDate = require('./fields/organisation-start-date');
 const fieldOrganisationType = require('./fields/organisation-type');
+const fieldProjectCountry = require('./fields/project-country');
 const fieldProjectTotalCosts = require('./fields/project-total-costs');
 const fieldSeniorContactRole = require('./fields/senior-contact-role');
 const fieldTotalIncomeYear = require('./fields/total-income-year');
@@ -585,48 +585,6 @@ module.exports = function fieldsFor({ locale, data = {} }) {
         };
     }
 
-    function fieldProjectCountry() {
-        const options = countriesFor({
-            locale: locale,
-            allowedCountries: config.get('awardsForAll.allowedCountries')
-        });
-
-        const activeOptions = options.filter(
-            option => has(option, 'attributes.disabled') === false
-        );
-
-        return {
-            name: 'projectCountry',
-            label: localise({
-                en: `What country will your project be based in?`,
-                cy: `Pa wlad fydd eich prosiect wedi’i leoli?`
-            }),
-            explanation: localise({
-                en: oneLine`We work slightly differently depending on which
-                    country your project is based in, to meet local needs
-                    and the regulations that apply there.`,
-                cy: oneLine`Rydym yn gweithredu ychydig yn wahanol, yn ddibynnol 
-                    ar pa wlad mae eich prosiect wedi’i leoli i ddiwallu 
-                    anghenion lleol a’r rheoliadau sy’n berthnasol yna.`
-            }),
-            type: 'radio',
-            options: options,
-            isRequired: true,
-            schema: Joi.string()
-                .valid(activeOptions.map(option => option.value))
-                .required(),
-            messages: [
-                {
-                    type: 'base',
-                    message: localise({
-                        en: 'Select a country',
-                        cy: 'Dewiswch wlad'
-                    })
-                }
-            ]
-        };
-    }
-
     function fieldCompanyNumber() {
         return {
             name: 'companyNumber',
@@ -794,7 +752,7 @@ module.exports = function fieldsFor({ locale, data = {} }) {
             ]
         },
         projectDateRange: fieldProjectDateRange(),
-        projectCountry: fieldProjectCountry(),
+        projectCountry: fieldProjectCountry(locale),
         projectLocation: {
             name: 'projectLocation',
             label: localise({
